requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
autopep8==2.3.1
Brotli==1.0.9
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
docopt==0.6.2
einops==0.8.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gmpy2==2.1.2
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.66.0
h5py==3.11.0
httplib2==0.22.0
huggingface-hub==0.24.6
HyperPyYAML==1.2.2
idna==3.7
importlib_metadata==8.4.0
importlib_resources==6.4.4
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
keras==3.5.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.6
llvmlite==0.43.0
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
mdurl==0.1.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
ml-dtypes==0.4.0
more-itertools==10.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mutagen==1.47.0
namex==0.0.8
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
openai-whisper==20231117
opt-einsum==3.3.0
optree==0.12.1
optuna==3.6.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pip==24.2
platformdirs==4.2.2
pooch==1.8.2
primePy==1.3
proto-plus==1.24.0
protobuf==4.25.4
pyannote.audio==3.3.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pycparser==2.22
pycryptodomex==3.20.0
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytorch-metric-learning==2.6.0
pytube==15.0.0
pytubefix==6.13.0
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.6.2
scikit-learn==1.5.1
scipy==1.13.1
semver==3.0.2
sentencepiece==0.2.0
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.4.0
speechbrain==1.0.0
SQLAlchemy==2.0.32
sympy==1.12
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tomli==2.0.1
torch==2.4.0
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.4.0
torchmetrics==1.4.1
torchvision==0.19.0
tqdm==4.66.5
triton==2.3.1
typer==0.12.5
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
websockets==13.0
Werkzeug==3.0.4
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
yt-dlp==2024.8.6
zipp==3.20.0
